Improve performance of named operations on RasterStack/Brick
It turns out that pulling a RasterLayer out of a RasterBrick using the
subset method is very time-consuming. A much faster better way is to use
getValuesBlock on the RasterStack or RasterBrick directly.

A bug in the raster package prevents using this improvement; a fix is
proposed at https://github.com/rspatial/raster/pull/134